Fix failing test after migrating to Gradle 2.9
The test was failing due to incorrect stubbing of a static method. Use
the expectation-based stubbing of static methods: this is the correct
way of doing according to the PowerMockito javadoc and it makes the code
clearer.

See http://static.javadoc.io/org.powermock/powermock-api-mockito/1.6.4/org/powermock/api/mockito/expectation/PowerMockitoStubber.html#when-java.lang.Class-