add "clean*" tasks as dependencies of task "clean"
Gradle has a rule that a task called "clean<NameOfTask>" clears the output
of the task "<nameOfTask>". However, by default these task-specific
cleaners are not executed when running "gradle clean". It doesn't matter
as long as a custom task's output directory is in the project's default
build directory, but if that's not the case, the outputs won't be
cleaned unless you specifically request it.

This commit makes it so that the task "clean", if defined in the project
at the moment of the javacc plugin application, is set to depend on the
cleaner counterparts of the tasks generated by this plugin. This way, even if
the generated source directory is moved out of the project's build
directory, it still gets cleaned when the user runs "gradle clean".