Remove build step for generating the README file
The current release process for the plugin includes a step which
regenerates the `README.md` file from the `buildSrc/README.md.template`
file and commits it to the repository.  This ensures that the README is
always updated with the latest version number, but it adds an extra step
and requires updates to the README info to be made in an unexpected
location.  This could be confusing, especially for new contributors who
are submitting a pull request for the first time.

From this point forward, updates to the README info should be made
directly in the root `README.md` file.  In the CHANGELOG section, the
in-progress release should be listed with a `-SNAPSHOT` version number.
It will be the responsibility of the maintainer to ensure that the file
contains the correct version numbers when publishing a release.