Update import order configurations
There is a standard `org` group for third-party imports, but that also
matches the new `org.javacc.plugin` package-name prefix for local
imports and results in an incorrect order.  The `org.javacc.plugin`
prefix needs to be added explicitly at the end of the group list to
force a separate group in the correct location (because longest match
wins).